insider_hiring_signals/requirements.txt
# Insider Activity + Hiring Signals System
# Python 3.11+ required

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
selenium==4.16.0
webdriver-manager==4.0.1

# SEC EDGAR Parsing
python-dateutil==2.8.2
xmltodict==0.13.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# AI/LLM Integration
anthropic==0.8.1
openai==1.6.1

# Email
python-dotenv==1.0.0

# Configuration
pyyaml==6.0.1

# Scheduling
schedule==1.2.1
apscheduler==3.10.4

# Rate Limiting
ratelimit==2.2.1
backoff==2.2.1

# Logging
loguru==0.7.2

# CLI and Reporting
rich==13.7.0
tabulate==0.9.0
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
responses==0.24.1

# Type Hints
typing-extensions==4.9.0
